[CHANGE] Optimized API & Moved to 2 Endpoints
