+++ madc/api/admin/__init__.py
@@ -1,0 +1,6 @@
+# pylint: disable=unused-argument
+from .admin import DoctrineCheckerAdminApiEndpoints
+
+
+def setup(api):
+    DoctrineCheckerAdminApiEndpoints(api)
